fix make error

+++ Makefile
@@ -21,7 +21,7 @@ build:
 	@echo "build"
 	go build -v -trimpath \
 		-o ${RELEASE_DIR}/cell ./cmd/cell
-	rm cell
+	test ! -f "cell" ||  rm cell
 	ln -s ${RELEASE_DIR}/cell cell
 test:
 	@echo "unit test"
